Add files via upload
With online sales gaining popularity, tech companies are exploring ways to improve their sales by analysing customer behaviour and gaining insights about product trends. Furthermore, the websites make it easier for customers to find the products they require without much scavenging. Needless to say, the role of big data analysts is among the most sought-after job profiles of this decade. 
For this assignment, WE will be working with a public clickstream dataset of a cosmetics store. Using this dataset, OURjob is to extract valuable insights which generally data engineers come up within an e-retail company.